+++ packages/ui/app/src/resolver/__test__/ApiDefinitionResolver.ts
@@ -3,7 +3,7 @@ import * as FernNavigation from "@fern-api/fdr-sdk/navigation";
 import fs from "fs";
 import path from "path";
 import { DEFAULT_FEATURE_FLAGS } from "../../atoms";
-import { ApiReferenceResolver } from "../ApiReferenceResolver";
+import { ApiDefinitionResolver } from "../ApiDefinitionResolver";
 import { ApiTypeResolver } from "../ApiTypeResolver";
 import { ResolvedEndpointDefinition } from "../types";
 
@@ -31,7 +31,7 @@ describe("resolveApiDefinition", () => {
             FernNavigation.SlugGenerator.init(""),
         );
 
-        const resolved = await ApiReferenceResolver.resolve(
+        const resolved = await ApiDefinitionResolver.resolve(
             node,
             holder,
             typeResolver,
@@ -65,7 +65,7 @@ describe("resolveApiDefinition", () => {
             FernNavigation.SlugGenerator.init(""),
         );
 
-        const resolved = await ApiReferenceResolver.resolve(
+        const resolved = await ApiDefinitionResolver.resolve(
             node,
             holder,
             typeResolver,
